+++ app/src/main/java/ua/com/radiokot/photoprism/features/gallery/GalleryFeatureModule.kt
@@ -139,6 +139,7 @@ val galleryFeatureModule = module {
                 backgroundMediaFileDownloadManager = get(),
                 downloadFileUseCase = get(),
                 downloadUrlFactory = get(),
+                galleryPreferences = get(),
             )
         } bind GalleryMediaDownloadActionsViewModelDelegate::class